refactor: adjust the code structure
as title

Log: adjust the code structure
