+++ src/Idle.h
@@ -8,27 +8,22 @@
 #include <QTimer>
 #include <QObject>
 
-#define DUM_AUTO_IDLE_TIMEOUT 60000
 
 class Idle: public QObject
 {
     Q_OBJECT
 public:
     explicit Idle(QObject *parent = nullptr);
+
     void Inhibit(const QString& task);
     void UnInhibit(const QString& task);
 
 private:
     void handleInhibit() const;
 
-public Q_SLOTS:
-    void onTimeout() const;
-
 private:
     QTimer *m_timer{};
     QStringList m_reasons;
 };
-
-
 
 #endif //IDLE_H
